refactor: remove bundler wrapper files and enable vitest typecheck
Remove vite, rolldown, esbuild, rollup, webpack wrappers — consumers
use logsSDK/logsSDKServer directly from the unplugin export. Split type
tests into diagnostics.test-d.ts and logger.test-d.ts, fix stale imports,
and enable vitest typecheck.
