--- src/code-transform/esbuild.ts
@@ -1,4 +1,0 @@
-import type { LogsSdkPluginOptions } from './unplugin'
-import { unplugin } from './unplugin'
-
-export default unplugin.esbuild as (options?: LogsSdkPluginOptions) => any
