--- src/code-transform/rolldown.ts
@@ -1,4 +1,0 @@
-import type { LogsSdkPluginOptions } from './unplugin'
-import { logsSDK } from './unplugin'
-
-export default logsSDK.rolldown as (options?: LogsSdkPluginOptions) => any
