--- src/code-transform/rollup.ts
@@ -1,4 +1,0 @@
-import type { LogsSdkPluginOptions } from './unplugin'
-import { unplugin } from './unplugin'
-
-export default unplugin.rollup as (options?: LogsSdkPluginOptions) => any
