--- src/code-transform/webpack.ts
@@ -1,4 +1,0 @@
-import type { LogsSdkPluginOptions } from './unplugin'
-import { unplugin } from './unplugin'
-
-export default unplugin.webpack as (options?: LogsSdkPluginOptions) => any
