+++ tsdown.config.ts
@@ -7,8 +7,6 @@ export default defineConfig({
     'formatters/ansi': 'src/formatters/ansi.ts',
     'formatters/json': 'src/formatters/json.ts',
     'unplugin': 'src/code-transform/unplugin.ts',
-    'vite': 'src/code-transform/vite.ts',
-    'rolldown': 'src/code-transform/rolldown.ts',
     'dev-reporter': 'src/dev-reporter.ts',
   },
   dts: {
